requirements.txt
Django == 4.1.5
djangorestframework   == 3.14.0
django-environ == 0.9.0
psycopg2  == 2.9.5
sqlparse == 0.4.3
django-environ == 0.9.0
djangorestframework-simplejwt == 5.2.2
drf-yasg == 1.21.5
minio == 7.1.13